Refactor interactive DAG visualizer into modular TUI structure
